// scripts/src/common/interfaces/tx.ts
export type TxResponse =
  | (GetTransactionApiResponseBase & {
      version?: TransactionVersion;
      meta:
        | (TransactionMetaBase &
            TransactionMetaInnerInstructionsNotParsed &
            TransactionMetaLoadedAddresses)
        | null;
      transaction: TransactionJson & TransactionAddressTableLookups;
    })
  | null;

type GetTransactionApiResponseBase = Readonly<{
  /**
   * The estimated production time at which the transaction was processed. `null` if not
   * available.
   */
  blockTime: UnixTimestamp | null;
  /** The slot during which this transaction was processed */
  slot: Slot;
}>;

type ReturnData = {
  /** A tuple whose first element is the bytes of the return data as a base64-encoded string. */
  data: Base64EncodedDataResponse;
  /** The address of the program that generated the return data */
  programId: Address;
};

type TransactionMetaBase = Readonly<{
  /** Number of compute units consumed by the transaction */
  computeUnitsConsumed?: bigint;
  /** Error if transaction failed, `null` if transaction succeeded. */
  err: TransactionError | null;
  /** The fee this transaction was charged, in {@link Lamports} */
  fee: Lamports;
  /**
   * String log messages or `null` if log message recording was not enabled during this
   * transaction
   */
  logMessages: readonly string[] | null;
  /** Account balances after the transaction was processed */
  postBalances: readonly Lamports[];
  /**
   * List of token balances from after the transaction was processed or omitted if token balance
   * recording was not yet enabled during this transaction
   */
  postTokenBalances?: readonly TokenBalance[];
  /** Account balances from before the transaction was processed */
  preBalances: readonly Lamports[];
  /**
   * List of token balances from before the transaction was processed or omitted if token balance
   * recording was not yet enabled during this transaction
   */
  preTokenBalances?: readonly TokenBalance[];
  /** The most-recent return data generated by an instruction in the transaction */
  returnData?: ReturnData;
  /**
   * Transaction-level rewards; currently only `"Rent"`, but other types may be added in the
   * future
   */
  rewards: readonly Reward[] | null;
  /** @deprecated */
  status: TransactionStatus;
}>;

type InnerInstructions<TInstructionType> = Readonly<{
  /** The index of the instruction in the transaction */
  index: number;
  /** The instructions */
  instructions: readonly TInstructionType[];
}>;
// ...
type TransactionInstruction = InstructionWithData<ReadonlyUint8Array> &
  Partial<InstructionWithStackHeight> &
  Readonly<{
    /**
     * An ordered list of indices that indicate which accounts in the transaction message's
     * accounts list are loaded by this instruction.
     */
    accounts: readonly number[];
    /**
     * The index of the address in the transaction message's accounts list associated with the
     * program to invoke.
     */
    programIdIndex: number;
  }>;

type TransactionMetaInnerInstructionsNotParsed = Readonly<{
  /** A list of instructions called by programs via cross-program invocation (CPI) */
  innerInstructions?:
    | readonly InnerInstructions<TransactionInstruction>[]
    | null;
}>;

type AddressTableLookup = Readonly<{
  /** The address of the address lookup table account. */
  accountKey: Address;
  /** Indexes of accounts in a lookup table to load as read-only. */
  readonlyIndexes: readonly number[];
  /** Indexes of accounts in a lookup table to load as writable. */
  writableIndexes: readonly number[];
}>;

type TransactionAddressTableLookups = Readonly<{
  message: Readonly<{
    /** A list of address tables and the accounts that this transaction loads from them */
    addressTableLookups: readonly AddressTableLookup[];
  }>;
}>;

type TransactionMetaLoadedAddresses = Readonly<{
  /** Addresses loaded from lookup tables */
  loadedAddresses: {
    /** Ordered list of base-58 encoded addresses for read-only accounts */
    readonly: readonly Address[];
    /** Ordered list of base-58 encoded addresses for writable accounts */
    writable: readonly Address[];
  };
}>;

type TransactionBase = Readonly<{
  message: {
    /**
     * For transactions whose lifetime is specified by a recent blockhash, this is that
     * blockhash, and for transactions whose lifetime is specified by a durable nonce, this is
     * the nonce value.
     */
    recentBlockhash: Blockhash;
  };
  /**
   * An ordered list of signatures belonging to the accounts required to sign this transaction.
   *
   * Each signature is an Ed25519 signature of the transaction message using the private key
   * associated with the account required to sign the transaction.
   */
  signatures: readonly Base58EncodedBytes[];
}>;

type TransactionJson = Readonly<{
  message: {
    /** An ordered list of addresses belonging to the accounts loaded by this transaction */
    accountKeys: readonly Address[];
    header: {
      /**
       * The number of read-only accounts in the static accounts list that must sign this
       * transaction.
       *
       * Subtracting this number from `numRequiredSignatures` yields the index of the first
       * read-only signer account in the static accounts list.
       */
      numReadonlySignedAccounts: number;
      /**
       * The number of accounts in the static accounts list that are neither writable nor
       * signers.
       *
       * Adding this number to `numRequiredSignatures` yields the index of the first read-only
       * non-signer account in the static accounts list.
       */
      numReadonlyUnsignedAccounts: number;
      /**
       * The number of accounts in the static accounts list that must sign this transaction.
       *
       * Subtracting `numReadonlySignedAccounts` from this number yields the number of
       * writable signer accounts in the static accounts list. Writable signer accounts always
       * begin at index zero in the static accounts list.
       *
       * This number itself is the index of the first non-signer account in the static
       * accounts list.
       */
      numRequiredSignatures: number;
    };
    instructions: readonly TransactionInstruction[];
  };
}> &
  TransactionBase;